Fix regression that sent some all-sky images to all NaNs
This fixes a regression that was introduced in reproject-0.3
by commit 59c5b493c7807f70b19f0e337a88a9e0f1a3eccf that caused
certain all-sky images (e.g. the Mellinger Milky Way Panorama,
http://www.milkywaysky.com) to be reprojected to all NaNs when
the output WCS was in Mollweide coordinates.

The regression was due to a change that converted the minima and
maxima of reprojected indices from floats to integers so that
Numpy would not complain about non-integer array indices. If any
of the reprojected indices were NaN, then the minima and maxima
would be NaN. Converting the indices to integers broke some all-
sky reprojections because some subsequent code was inadvertently
relying on the fact that comparisons between NaNs and other
floating point values are always false.

The workaround is to find the minima and maxima of the non-NaN
values of the reprojected indices using `np.nanmin` and
`np.nanmax` respectively.