docs/mosaicking.rst: fix test case
The failing unicode-to-bytes comparison causes the list of HDUs to be
empty. Note that this caused failing behavior before the change to
`parse_input_shape` as well. I think this start failing with the recent
release of Astropy 4.1, and I think it's just a coincidence that this
failure affects the API being modified in this pull request.