Add support for un-centered Jacobians, make default
This reduces the number of coordinate transformations that need to be
done by 2/3. In my test configuration, this is worthwhile speed increase
with a vanishingly small loss in accuracy. For the test case with an AIA
image, the largest few per-pixel changes are a few percent where the
transformation is most extreme, while 99.6% of non-NaN pixels change by
less than 1%, and 97% by less than 0.1%. This loss of accuracy should be
small for any case where the transformation varies smoothly between
output pixels, which should be the case for many common real-world
transformations.