Reduce size of sampled input region to only what's needed
This is in two parts:

For each output pixel, a region of the input image is iterated through.
The input pixels of interest are, in general, a parallelogram at an
arbitrary orientation angle, and so a rectangular region surrounding
this parallelogram is scanned through. This update reduces the size of
that rectangular region to the smallest bounding box containing all the
pixels of interest. This comes with significant speed gains.

Second, as each row of the bounding box is scanned along, we now end the
iteration once we've scanned into and out of the parallelogram, skipping
any consideration of the remaining pixels in the row. This is a small
but simple gain in speed.