Optimize out-of-bounds cases
When an output image pixel is sampling input-image coordinates that
actually fall outside the input image, we can detect that early and do
less work for those cases. This offers significant speedups for
transformations where this case is common (e.g. a 45-degree rotation,
where the corners of the output image map to out-of-bounds input
coordinates).