Fix Jacobian calculation in the adaptive algorithm
The Jacobian matrix Ji was being calculated as the transpose of what it
should be, if it's to be used as a standard transformation-as-a-matrix.
Its inverse, J, was then also transposed. When this J is eventually
used, it was used in the transform-as-a-matrix manner.

The effect of this transposition appears to have been minor. If the
Jacobian matrix is decomposed into a scaling, a shearing, and a
rotation, the scaling is unaffected, x and y shearings are swapped, and
the rotation's direction is reversed. When thinking of circles in the
output plane being mapped to ellipses in the input plane, within which
to sample pixels to form the output pixel value, those ellipses would be
oriented differently than they should be, but they would have the same
overall area.

For the tests currently implemented, the differences in the output
images before and after this fix tend to be small and are not visually
apparent.