Remove incorrect normalization in Hann window
The window should carry a factor of 1/2 for each of the two cosine terms
being multiplied together, for a combined factor of 1/4. However, the
normalization factor comes out anyway when output pixel values are
divided by the sum of the weights, so it's a micro-optimization to just
remove the factor rather than correcting it.