Use padded singular values for building J matrix
The matrix J was being built using the un-padded values, whereas the
padded values ought to be used (as called for by DeForest (2004)). This
caused incorrect averaging weights to be computed for some sample
values, in transformations where the padding of singular values is
required.

For the implemented test cases, the effect of this fix is quite small.